sparkline: add sparklineGraph to do axis/label drawing
The idea here is to leave the sparkline function largely untouched, and wrap up
the sparkline drawn in some axes and axis labels to make it look more like a
graph.

Currently doesn't quite work -- I don't really understand how the sparkline
animation is interacting with the clippath I've added to restrict the drawing of
the data line, but somehow the first piece of the data isn't being drawn.

Also the x-axis label should probably be a bit more intelligent...